move tuple slice tests to slice_test.py
